backend/requirements.txt
Django>=6.0.0
djangorestframework>=3.15.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
transformers>=4.40.0
torch==2.9.0
openai>=1.0.0
psycopg2-binary>=2.9.9

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
coverage>=7.3.0

# Linting and formatting
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
django-stubs>=4.2.0
